+++ site/dat/docs/Changelog.md
@@ -4,6 +4,7 @@
  - load settings applied to Stepping Thread Groups
  - shellexec service: print to stdout/stderr when `out`/`err` options set to `null`
  - JMeter: append `user.properties` instead of overriding
+ - move pass/fail to services
 
 ## 0.4.4
  - limit max size of file that goes into zip for artifact upload (max-size option)