+++ site/dat/docs/Changelog.md
@@ -5,6 +5,7 @@
  - shellexec service: print to stdout/stderr when `out`/`err` options set to `null`
  - JMeter: append `user.properties` instead of overriding
  - move pass/fail to services
+ - set BlazeMeter session note with test error message if present
 
 ## 0.4.4
  - limit max size of file that goes into zip for artifact upload (max-size option)